Add hook to allow manifest update to be triggered externally
This allow the plugin to generate snapshot manifest even when the
underlying git repositories were updated outside of Gerrit.

(Example: when a project hosted by Gerrit is a mirror of an upstream
project and it is periodically fetch by an external script.  RefUpdated
event is not available from Gerrit in this case.)

Change-Id: Ie119ad393fd51d6aba4679973c0b7d9db3fbde47